requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
numpy
streamlit
matplotlib
